поправлен нейминг, в топлокациях убраны лишние стили
