+++ src/styles/main/Main.module.scss
@@ -93,14 +93,6 @@
       width: 100%;
       height: 100%;
     }
-
-    &:hover {
-      //box-shadow: 0 0.125rem 0.125rem -0.125rem rgba(31, 27, 45, 0.08), 0 0.25rem 0.75rem rgba(31, 27, 45, 0.08);
-
-      & .locations__icon_wrapper {
-        opacity: 1;
-      }
-    }
   }
 
   &__overlay {
@@ -127,7 +119,6 @@
       position: absolute;
       top: 0;
       right: 0;
-      opacity: 0;
       transition: $transition;
     }
   }
